Move validation to validation service package
